Added empty state to Pie Chart Renamed from interactive to BA Removed unnused file
