+++ dashboard/src/app/dashboard/[dashboardId]/devices/DevicesChartsSection.tsx
@@ -3,7 +3,7 @@
 import { use } from 'react';
 import DeviceUsageTrendChart from '@/app/dashboard/[dashboardId]/devices/DeviceUsageTrendChart';
 import { fetchDeviceTypeBreakdownAction, fetchDeviceUsageTrendAction } from '@/app/actions';
-import InteractivePieChart from '@/components/InteractivePieChart';
+import BAPieChart from '@/components/BAPieChart';
 import { getDeviceColor, getDeviceLabel } from '@/constants/deviceTypes';
 import { DeviceIcon } from '@/components/icons';
 
@@ -24,7 +24,7 @@ export default function DevicesChartsSection({
       <div className='bg-card border-border rounded-lg border p-6 shadow'>
         <h2 className='text-foreground mb-1 text-lg font-bold'>Device Types</h2>
         <p className='text-muted-foreground mb-4 text-sm'>Visitors by device category</p>
-        <InteractivePieChart
+        <BAPieChart
           data={deviceBreakdown}
           getColor={getDeviceColor}
           getLabel={getDeviceLabel}
